# docs/swagger/loyaltysystem-api-docs.yaml
openapi: 3.0.3
info:
  title: Loyalty System API
  description: |
    This is the API documentation for the Loyalty System platform, covering all the endpoints required for:
    * Customer App - Mobile application for end customers
    * Staff App - Application for store staff to handle in-store operations
    * Admin App - Administrative portal for business owners and managers
  version: 1.0.0
servers:
  - url: https://api.loyaltysystem.com/v1
    description: Production server
  - url: https://staging-api.loyaltysystem.com/v1
    description: Staging server
  - url: http://localhost:5000
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - confirmPassword
        - firstName
        - lastName
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        confirmPassword:
          type: string
          format: password
        firstName:
          type: string
        lastName:
          type: string
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/User'
        expiration:
          type: string
          format: date-time
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    
    # Loyalty Program schemas
...
    LoyaltyCard:
      type: object
      properties:
        id:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        loyaltyProgramId:
          type: string
          format: uuid
        cardNumber:
          type: string
        currentPoints:
          type: integer
        createdAt:
          type: string
          format: date-time
        lastUsed:
          type: string
          format: date-time
        status:
          type: string
          enum: [Active, Suspended, Expired]
    CreateCardRequest:
      type: object
      required:
        - loyaltyProgramId
      properties:
        loyaltyProgramId:
          type: string
          format: uuid

    # Reward schemas
...
    Error:
      type: object
      properties:
        message:
          type: string
        details:
          type: string
security:
  - bearerAuth: []
paths:
...
  /api/auth/register:
    post:
      tags:
        - Customer Authentication
      summary: Register a new customer account
      description: Creates a new customer user account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Successfully registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad request, validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/login:
    post:
      tags:
        - Customer Authentication
      summary: Login to customer account
      description: Authenticates a customer user and returns JWT token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successfully authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/loyalty-summary:
    get:
      tags:
        - Customer Authentication
      summary: Get customer loyalty summary
      description: Returns a summary of the customer's loyalty status
      responses:
        '200':
          description: Successfully retrieved loyalty summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  customerId:
                    type: string
                  activeProgramCount:
                    type: integer
                  totalPoints:
                    type: integer
                  totalStamps:
                    type: integer
                  pendingRewards:
                    type: integer
        '401':
          description: Unauthorized
        '400':
          description: Bad request

  # Loyalty Programs (Customer)
...
  /api/loyaltyCards:
    post:
      tags:
        - Loyalty Cards
      summary: Enroll in a loyalty program
      description: Creates a new loyalty card for the customer in a specific program
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCardRequest'
      responses:
        '201':
          description: Successfully created loyalty card
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoyaltyCard'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
...
  /api/loyaltyCards/{id}:
    get:
      tags:
        - Loyalty Cards
      summary: Get loyalty card by ID
      description: Returns details of a specific loyalty card
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved loyalty card
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoyaltyCard'
        '404':
          description: Card not found
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - card belongs to another customer
...
  /api/admin/auth/login:
    post:
      tags:
        - Admin Authentication
      summary: Login to admin account
      description: Authenticates an admin user and returns JWT token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successfully authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Authentication failed

  # Business Management (Admin)
...
  /api/staff/auth/login:
    post:
      tags:
        - Staff Authentication
      summary: Login to staff account
      description: Authenticates a staff user and returns JWT token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successfully authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Authentication failed

  # Customer Management (Staff)
...